+++ src/components/bannerLower/index.tsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-
 import style from './bannerLower.module.scss';
 
 export default class BannerLower extends Component<any> {
